refactor: split nearby presence handlers
