+++ scripts/ios_pr_check.sh
@@ -44,6 +44,7 @@ swift scripts/backend_contract_versioning_unit_check.swift
 swift scripts/backend_edge_observability_unit_check.swift
 swift scripts/backend_edge_auth_unification_unit_check.swift
 swift scripts/sync_walk_stage_handler_split_unit_check.swift
+swift scripts/nearby_presence_handler_split_unit_check.swift
 swift scripts/rival_privacy_policy_stage1_unit_check.swift
 swift scripts/rival_privacy_policy_confirmed_unit_check.swift
 swift scripts/rival_privacy_hard_guard_unit_check.swift
