+++ supabase/functions/nearby-presence/support/core.ts
@@ -1,0 +1,85 @@
+export const json = (body: unknown, status = 200) =>
+  new Response(JSON.stringify(body), {
+    status,
+    headers: { "Content-Type": "application/json" },
+  });
+
+export const roundCoord = (value: number) => Math.round(value * 10_000) / 10_000;
+
+export const asUUIDOrNull = (value: unknown): string | null => {
+  if (typeof value !== "string") return null;
+  const normalized = value.trim().toLowerCase();
+  const uuidPattern = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/;
+  return uuidPattern.test(normalized) ? normalized : null;
+};
+
+export const asFiniteNumberOrNull = (value: unknown): number | null => {
+  if (typeof value !== "number") return null;
+  return Number.isFinite(value) ? value : null;
+};
+
+export const asISO8601OrNull = (value: unknown): string | null => {
+  if (typeof value !== "string") return null;
+  const parsed = Date.parse(value);
+  if (Number.isNaN(parsed)) return null;
+  return new Date(parsed).toISOString();
+};
+
+export const asPositiveIntegerOrNull = (value: unknown): number | null => {
+  if (typeof value !== "number" || Number.isFinite(value) === false) return null;
+  const normalized = Math.floor(value);
+  return normalized > 0 ? normalized : null;
+};
+
+export const asNonEmptyTextOrNull = (value: unknown): string | null => {
+  if (typeof value !== "string") return null;
+  const normalized = value.trim();
+  return normalized.length > 0 ? normalized : null;
+};
+
+export const asUUIDArray = (value: unknown): string[] => {
+  if (Array.isArray(value) === false) return [];
+  return value
+    .map((entry) => asUUIDOrNull(entry))
+    .filter((entry: string | null): entry is string => entry != null);
+};
+
+export const geohashEncode = (lat: number, lng: number, precision = 7): string => {
+  const base32 = "0123456789bcdefghjkmnpqrstuvwxyz";
+  const bits = [16, 8, 4, 2, 1];
+  let latRange: [number, number] = [-90, 90];
+  let lngRange: [number, number] = [-180, 180];
+  let hash = "";
+  let bit = 0;
+  let ch = 0;
+  let even = true;
+
+  while (hash.length < precision) {
+    if (even) {
+      const mid = (lngRange[0] + lngRange[1]) / 2;
+      if (lng >= mid) {
+        ch |= bits[bit];
+        lngRange = [mid, lngRange[1]];
+      } else {
+        lngRange = [lngRange[0], mid];
+      }
+    } else {
+      const mid = (latRange[0] + latRange[1]) / 2;
+      if (lat >= mid) {
+        ch |= bits[bit];
+        latRange = [mid, latRange[1]];
+      } else {
+        latRange = [latRange[0], mid];
+      }
+    }
+    even = !even;
+    if (bit < 4) {
+      bit += 1;
+    } else {
+      hash += base32[ch];
+      bit = 0;
+      ch = 0;
+    }
+  }
+  return hash;
+};
